requirements.txt
psycopg2-binary==2.8.4
pymongo==3.9.0
dnspython==1.16.0
bottle==0.12.17
requests==2.21.0
nltk==3.4.5
python-dotenv==0.10.3
jsonschema==2.6.0